extracted iterated filtering into new file
