+++ src/probnum/filtsmooth/gaussfiltsmooth/__init__.py
@@ -1,4 +1,5 @@
 from .extendedkalman import *
+from .iterated_kalman import *
 from .kalman import *
 from .kalmanposterior import *
 from .stoppingcriterion import *
